Feat(sequelize): add sequelize and initialize postgres connection
Description:
In this project, it is decided to use the postgres database as the main database.
Therefore, in this commit, I added the sequelize configuration file, which is an
ORM for SQL databases, to the system. I also created a connection between the
project and the postgres database through this module.

Change Logs:
[Packages] [Added] - add sequelize and postgres related modules.
[Sequelize config] [Added] - Add sequelize module config.
[README] [Added] - Add a basic readme file.